fix: correct handling of polarizations from UVBeam
This commit updates the behaviour of vis_cpu to be more consistent.
In particular, when ``polarized=False``, we explicitly take the sqrt
of the power beam for a particular feed. We raise errors for incompatible
inputs.

BREAKING CHANGE: previous inputs for the beam will now sometimes
raise errors, as they were inconsistent. In particular, input
``beam_list`` to ``vis_cpu`` must be efield type if using `polarized=True`
and `power` type if not.