test: update test against pyuvsim
This test uses a small, downsampled version of the Nic Fagnoni beam
models for the HERA antennas outfitted with the PAPER dipole.
Interestingly, the test for the polarized case passes with the original
tolerances, but the test for the unpolarized case has up to ~percent
level differences in the simulated visibilities. It is not clear at this
time why the discrepancy between the simulations is greater for the
unpolarized case.

This commit also removes the silent change to the beam normalization.